+++ test/Fixture/Miscellaneous/ExampleTraitWithSecondProtectedMethod.php
@@ -1,0 +1,27 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Copyright (c) 2023-2026 Andreas Möller
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE.md file that was distributed with this source code.
+ *
+ * @see https://github.com/ergebnis/rector-rules
+ */
+
+/**
+ * Copyright (c) 2023-2026 Andreas Möller.
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE.md file that was distributed with this source code.
+ *
+ * @see https://github.com/ergebnis/rector-rules
+ */
+trait ExampleTraitWithSecondProtectedMethod
+{
+    protected function secondProtectedMethod(): void
+    {
+    }
+}