+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 For a full diff see [`1.12.0...main`][1.12.0...main].
 
+### Added
+
+- Added `Rules\Traits\RemoveUnusedTraitUseRector` to remove unused trait `use` statements from `final` classes ([#298]), by [@localheinz]
+
 ## [`1.12.0`][1.12.0]
 
 For a full diff see [`1.11.1...1.12.0`][1.11.1...1.12.0].
@@ -268,5 +272,6 @@ For a full diff see [`fd198f0...0.1.0`][fd198f0...0.1.0].
 [#295]: https://github.com/ergebnis/rector-rules/pull/295
 [#296]: https://github.com/ergebnis/rector-rules/pull/296
 [#297]: https://github.com/ergebnis/rector-rules/pull/297
+[#298]: https://github.com/ergebnis/rector-rules/pull/298
 
 [@localheinz]: https://github.com/localheinz